# auth/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AuthBase API
  version: 1.0.0
  description: |
    AuthBase の認証APIです。

    ## 認証方式

    エンドユーザー向けエンドポイント（`/me`, `/token`, `/logout`, `/bridge/issue` など）は
    `Authorization` ヘッダーにセッショントークン（JWT/HS512）を渡して認証します。

    管理者向けエンドポイント（`/api/*`）はブラウザのCookieセッションで認証します（`/admin/login` でログイン）。
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
...
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      description: |
        `/basic/signup` または `/basic/login` / OAuthコールバック で取得したセッショントークン（JWT/HS512）。
        `Authorization: Bearer <token>` の形式で送信します。
    AdminCookieSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        `/admin/login` で発行されるCookieセッション。
        ブラウザでは自動送信されます。

  # ------------------------------------------------------------
  # Schemas
  # ------------------------------------------------------------
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "unauthorized"
...
    PublicUserInfo:
      type: object
      description: 公開ユーザー情報（`/info/:userid` レスポンス）
      properties:
        user_id:
          type: string
          example: "01HXXXXXXXXXXXXXXXXXXXXXXXXX"
        name:
          type: string
          example: "山田太郎"
...
paths:
...
  /info/{userid}:
    get:
      summary: ユーザーの公開情報を取得
      description: ユーザーIDから公開情報（ID・名前のみ）を取得します。認証不要。
      operationId: getUserInfo
      tags: [User]
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: string
          example: "01HXXXXXXXXXXXXXXXXXXXXXXXXX"
      responses:
        "200":
          description: 公開ユーザー情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUserInfo"
        "500":
          description: ユーザーが見つからない、またはサーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ----------------------------------------------------------
  # アクセストークンからユーザー詳細を取得
  # ----------------------------------------------------------
...